restructured directory layout so lazy can detect changes and auto reload
