+++ init.lua
@@ -1,2 +1,3 @@
-print("hello")
 require('CYB3RKUN')
+require('lazy_init')
+
