+++ lazy_init.lua
@@ -1,3 +1,4 @@
+
 local lazypath = vim.fn.stdpath('data') .. '/lazy/lazy.nvim'
 if not vim.loop.fs_stat(lazypath) then
 	vim.fn.system({
@@ -11,7 +12,5 @@ if not vim.loop.fs_stat(lazypath) then
 end
 vim.opt.rtp:prepend(lazypath)
 
-require('lazy').setup({
-	spec = 'CYB3RKUN.plugins',
-	change_detection = {notify = false}
-})
+require('lazy').setup("plugins")
+
