+++ lua/plugins/autopairs.lua
@@ -6,3 +6,4 @@ return {
 		--use opts = {} for passing setup options
 }
 
+
