+++ lua/plugins/surround.lua
@@ -6,3 +6,4 @@ return {
         require("nvim-surround").setup({})
     end
 }
+
