+++ lua/plugins/todo.lua
@@ -11,3 +11,4 @@ return {
 		},
 	}
 }
+
