+++ lua/plugins/treesitter.lua
@@ -27,3 +27,4 @@ return {
 		})
 	end
 }
+
